+++ robot_descriptions/_descriptions.py
@@ -121,4 +121,5 @@ DESCRIPTIONS: Dict[str, Description] = {
     "valkyrie_description": Description(Format.URDF),
     "yumi_description": Description(Format.URDF),
     "z1_description": Description(Format.URDF),
+    "draco3_description": Description(Format.URDF),
 }